Updated colorsscatter.py
So I have amended this code to now suit the post_hess changes where I had created a function to not only record eigvecs for filaments but for sheets as well. Also and more to the importance of this change, I have also created a code within post_hess to create a mask which identified each pixel in the simulation to be 0-void,1-sheet,2-filament and 3-cluster. So I am using the mask in this code to identify halos that fall within appropriate spatial bins