New inclusions amid post_hess changes
Amid major changes to post_hess to include ALL structure identification, not just filaments, I have trickled down necessery inclusions such as 'mask' is a new array which is outputted from post_hess in order to indicate the LSS type for each pixel in the sim. Also, the eigvecs are filament and sheet axis.