Major changes
Instead of merely binning the angular momentum, I have also anticipated the use of velocity, virial mass and radius. So I have included the changes to bin all of these across the entire simulation. 
I have also preserved the 'resid' concept as it is very imporant there are no overlapping information, and there is no simpler way to do it